Restructure security groups util code
