+++ src/utils/security-groups/helpers.ts
@@ -1,5 +1,5 @@
 import { Peer } from "@aws-cdk/aws-ec2";
-import type { CidrIngress } from "../constructs/ec2";
+import type { CidrIngress } from "../../constructs/ec2";
 
 export const transformToCidrIngress = (ingresses: Array<[string, string]>): CidrIngress[] => {
   return ingresses.map(([key, value]) => {
